chore: 接入 ruff 与 mypy 配置并修复静态检查问题
pyproject 增加 ruff（E/F/I/UP/B/SIM/C4/DTZ/ISC/PIE/RUF，line-length 88）
和 mypy 宽松档配置。ruff --fix 处理 155 处导入排序、datetime.UTC、
过时导入和未用导入，其余 20 处手工处理。

顺带修复两个真实缺陷：
- agent_run_manager._finish_error 只拦 CANCELLED，已 COMPLETED/REJECTED
  的 run 会被后续异常覆写状态并追加重复终态事件，改为终态幂等；
- api.py 两处 except (ValueError, KeyError) 中的 KeyError 永不可达，
  因为前面已单独 except KeyError。

RUF001-003（中文全角标点）、UP042（(str, Enum) 是对外契约）、B905、
SIM103（guard-clause 风格）按项目实际情况关闭并注明原因。

check.ps1 在锁定引擎下 191 passed。

+++ command_center/masp/task_stream.py
@@ -1,11 +1,11 @@
 from __future__ import annotations
 
 import json
-import math
 import random
+from collections.abc import Mapping, Sequence
 from copy import deepcopy
 from pathlib import Path
-from typing import Any, Mapping, Sequence
+from typing import Any
 
 import networkx as nx
 from jsonschema import Draft202012Validator
@@ -127,7 +127,7 @@ def _poisson_arrivals(
     current = float(start_ms)
     while len(arrivals) < limit:
         current += rng.expovariate(1.0 / mean_interval_ms)
-        rounded = int(round(current))
+        rounded = round(current)
         if rounded >= end_ms:
             break
         arrivals.append(max(start_ms, rounded))
